safari-534.53-branch/LayoutTests/animations/animation-api-1-expected.txt

This tests the WebKitAnimation interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.WebKitAnimation is defined.
PASS window.WebKitAnimationList is defined.

Test an almost-default animation
PASS a_animations.length is 1
PASS a_animation_0 is defined.

PASS a_animation_0.name is 'test1'
PASS a_animation_0.duration is 1
PASS a_animation_0.delay is 0
PASS a_animation_0.iterationCount is 1
PASS a_animation_0.paused is false
PASS a_animation_0.ended is false
PASS a_animation_0.direction is window.WebKitAnimation.DIRECTION_NORMAL
PASS a_animation_0.fillMode is window.WebKitAnimation.FILL_NONE

Test a specified animation
PASS b_animations.length is 1
PASS b_animation_0 is defined.

PASS b_animation_0.name is 'test1'
PASS b_animation_0.duration is 2
PASS b_animation_0.delay is 3
PASS b_animation_0.iterationCount is 4
PASS b_animation_0.paused is false
PASS b_animation_0.ended is false
PASS b_animation_0.direction is window.WebKitAnimation.DIRECTION_NORMAL
PASS b_animation_0.fillMode is window.WebKitAnimation.FILL_NONE

Test negative delay, fill mode, direction and infinite iterations
PASS c_animations.length is 1
PASS c_animation_0 is defined.

PASS c_animation_0.name is 'test2'
PASS c_animation_0.duration is 5
PASS c_animation_0.delay is -1
PASS c_animation_0.iterationCount is Number.POSITIVE_INFINITY
PASS c_animation_0.paused is false
PASS c_animation_0.ended is false
PASS c_animation_0.direction is window.WebKitAnimation.DIRECTION_ALTERNATE
PASS c_animation_0.fillMode is window.WebKitAnimation.FILL_FORWARDS

Test multiple animations
PASS d_animations.length is 2
PASS d_animation_0 is defined.

PASS d_animation_0.name is 'test1'
PASS d_animation_0.duration is 1
PASS d_animation_0.delay is 0
PASS d_animation_0.iterationCount is 10
PASS d_animation_0.paused is false
PASS d_animation_0.ended is false
PASS d_animation_0.direction is window.WebKitAnimation.DIRECTION_NORMAL
PASS d_animation_0.fillMode is window.WebKitAnimation.FILL_BACKWARDS
PASS d_animation_1 is defined.

PASS d_animation_1.name is 'test2'
PASS d_animation_1.duration is 2
PASS d_animation_1.delay is 1
PASS d_animation_1.iterationCount is Number.POSITIVE_INFINITY
PASS d_animation_1.paused is false
PASS d_animation_1.ended is false
PASS d_animation_1.direction is window.WebKitAnimation.DIRECTION_ALTERNATE
PASS d_animation_1.fillMode is window.WebKitAnimation.FILL_BOTH

Test element with no animations
PASS e_animations is null

PASS successfullyParsed is true

TEST COMPLETE
